fix error with sonarcube
